# swagger-ui/swagger.yml
openapi: 3.0.0
info:
  title: ft_transcendance API
  description: |-
    ft_transcendance API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: /api
...
paths:
...
  /users/me/match_history:
    get:
      tags:
        - Matchs
      summary: Get the match history for the connected user
      security:
        - 42oauth: [public]
        - bearerAuth: []
      operationId: GetMatchHistory
      parameters:
        - $ref: "#/components/parameters/Onset"
        - $ref: "#/components/parameters/Length"
      responses:
        "200":
          $ref: "#/components/responses/MatchHistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  responses:
...
    Unauthorized:
      description: No token provided or invalid token provided
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorOutput"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorOutput"
...
    MatchHistoryResponse:
      description: success response
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/MatchOutput"
  parameters:
    Onset:
      description: "beginning index of the list (default: 0)"
      in: query
      name: onset
      required: false
      schema:
        type: integer
        format: int64
    Length:
      description: "size of the list (default: all)"
      in: query
      name: length
      required: false
      schema:
        type: integer
        format: int64
...
  schemas:
...
    ErrorOutput:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: some message explaining the error
    PlayerOutput:
      type: object
      required:
        - user
        - score
      properties:
        user:
          $ref: "#/components/schemas/UserOutput"
        score:
          type: integer
          format: int64
          example: 6
    MatchOutput:
      type: object
      required:
        - id
        - created_at
        - finished_at
        - loser
        - winner
      properties:
        id:
          type: integer
          format: int64
          example: 111
        created_at:
          type: string
          format: datetime
        finished_at:
          type: string
          format: datetime
        winner:
          $ref: "#/components/schemas/PlayerOutput"
        loser:
          $ref: "#/components/schemas/PlayerOutput"
    UserStatsOutput:
      type: object
      required:
        - id
        - level
        - victories
        - defeats
      properties:
        id:
          type: integer
          format: int64
          example: 11
        level:
          type: integer
          format: int64
          example: 21
        victories:
          type: integer
          format: int64
          example: 20
        defeats:
          type: integer
          format: int64
          example: 0
    AvatarOutput:
      type: object
      required:
        - id
        - path
      properties:
        id:
          type: integer
          format: int64
          example: 11
        path:
          type: string
          example: "/pictures/me.png"
    UserOutput:
      type: object
      required:
        - login
        - username
        - status
        - actual_avatar
        - stats
        - settings
      properties:
        login:
          type: string
          example: dboyer
        actual_avatar:
          $ref: "#/components/schemas/AvatarOutput"
        username:
          type: string
          example: dboyer
        status:
          type: string
          enum:
            - online
            - offline
            - in_game
            - in_queue
            - afk
        stats:
          $ref: "#/components/schemas/UserStatsOutput"
        settings:
          $ref: "#/components/schemas/UserSettings"
    UserSettings:
      type: object
      required:
        - two_fa
        - paddle_id
        - map_id
      properties:
        two_fa:
          type: boolean
          example: false
        paddle_id:
          type: integer
          enum:
            - 1
            - 2
            - 3
            - 4
        map_id:
          type: integer
          enum:
            - 1
            - 2
            - 3
            - 4
...
  securitySchemes:
    42oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: "https://api.intra.42.fr/oauth/authorize"
          tokenUrl: /api/auth/callback
          refreshUrl: /api/token
          scopes:
            public: Grant access to 42 public data
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT